[WIP]: Add redis support (#154)
* Add InMemoryBalanceTracker as a fallback for testing purposes.
* Improve test coverage
* Cleanup dependencies
* Formatting and docs

+++ ilpv4-connector-accounts/src/main/java/org/interledger/connector/accounts/AccountBalanceSettings.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import org.immutables.value.Value;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 /**
@@ -22,37 +21,52 @@ public interface AccountBalanceSettings {
   }
 
   /**
-   * The minimum balance (in this account's indivisible base units) the connector must maintain. The connector will
-   * reject outgoing packets if they would put it below this balance.
+   * <p>The minimum balance (in this account's indivisible base units) the connector must maintain for this account.
+   * For example, the connector will reject incoming packets if the transaction would put the account balance below this
+   * number. If this value is not present, then the connector will assume no minimum balance.</p>
+   *
+   * <p>Note that it is permissible to use a {@link Long} here since this value represents a minimum value, which is
+   * sign-less. Thus, we can use the largest long to represent the largest minimum balance.</p>
    *
    * @return The minimum balance, or {@link Optional#empty()} if there is no minimum.
    */
-  Optional<BigInteger> getMinBalance();
+  Optional<Long> getMinBalance();
 
   /**
-   * Maximum balance (in this account's indivisible base units) the connector will allow. The connector will reject
-   * incoming packets if they would put it above this balance. If this value is not present, then the connector will
-   * assume no maximum balance.
+   * <p>Maximum balance (in this account's indivisible base units) the connector will allow. For example, the connector
+   * will reject outgoing packets if the transaction would put the account balance above this number. If this value is
+   * not present, then the connector will assume no maximum balance.</p>
+   *
+   * <p>Note that it is permissible to use a {@link Long} here since this value represents a minimum value, which is
+   * sign-less. Thus, we can use the largest long to represent the largest minimum balance.</p>
    *
    * @return The maximum balance, or {@link Optional#empty()} if there is no maximum.
    */
-  Optional<BigInteger> getMaxBalance();
+  Optional<Long> getMaxBalance();
 
   /**
-   * Optional Balance (in this account's indivisible base units) numerically below which the connector will
-   * automatically initiate a settlement.
+   * <p>Optional Balance (in this account's indivisible base units) numerically below which the connector will
+   * automatically initiate a settlement.</p>
+   *
+   * <p>Note that it is permissible to use a {@link Long} here since this value represents a threshold for a value
+   * that will never go negative (this roughly correlates to the `prepaid_balance` in the balance tracker, and this will
+   * never go negative). Thus, we can use the largest long number to represent the largest threshold).</p>
    *
    * @return The settlement threshold balance, or {@link Optional#empty()} if there is no threshold (i.e., the account
    * should never settle).
    */
-  Optional<BigInteger> getSettleThreshold();
+  Optional<Long> getSettleThreshold();
 
   /**
    * Optional Balance (in this account's indivisible base units) the connector will attempt to reach when settling.
    *
+   * <p>Note that it is permissible to use a {@link Long} here since this value only ever represents a positive
+   * number (this roughly correlates to the `prepaid_balance` in the balance tracker, and this will never go negative).
+   * Thus, we can use the largest long number to represent the largest threshold).</p>
+   *
    * @return The amount that triggers settlement.
    */
-  Optional<BigInteger> getSettleTo();
+  Optional<Long> getSettleTo();
 
   @Value.Immutable(intern = true)
   @Value.Modifiable
